requirements.txt
discord.py>=2.0.0
python-dotenv
aiohttp
beautifulsoup4
pytz 